// desktop/frontend/packages/core/src/types/types.ts
export interface SaveData {
  val: any
  version: number
  updated_at?: string
}

export interface Snapshot {
  meta: {
    currentHash: string
    previousHash: string
    createdAt: number
  }
  data: {
    dict: string
    setting: string
    [PRACTICE_WORD_CACHE.key]: string
    [PRACTICE_ARTICLE_CACHE.key]: string
    [APP_VERSION.key]: number
  }
}

export interface BackupData {
  version: number
  val: {
    dict: SaveData
    setting: SaveData
    [PRACTICE_WORD_CACHE.key]: SaveData
    [PRACTICE_ARTICLE_CACHE.key]: SaveData
    [APP_VERSION.key]: number
  }
}